BaseContainer.scss: move to component
