+++ src/components/ui/BaseContainer.scss
@@ -1,4 +1,4 @@
-@import '../theme';
+@import 'src/styles/theme';
 
 /*
 * More information about flexbox under https://css-tricks.com/snippets/css/a-guide-to-flexbox/#aa-flexbox-properties
