+++ src/components/ui/BaseContainer.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import "../../styles/ui/BaseContainer.scss"
+import "./BaseContainer.scss"
 import PropTypes from "prop-types"
 
 const BaseContainer = (props) => (
